requirements.txt
aiohttp==3.7.2
arrow==0.17.0
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
base58==2.0.1
binance-chain==1.0.0
bitarray==1.2.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==6.6
cytoolz==0.11.0
Django==3.1.3
docker-py==1.8.0
ecdsa==0.16.1
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
flake8==3.8.4
funcsigs==1.0.2
future==0.18.2
hexbytes==0.2.1
idna==2.10
iniconfig==1.1.1
ipfshttpclient==0.7.0a1
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
jsonschema==3.2.0
lru-dict==1.1.6
mccabe==0.6.1
mnemonic==0.19
multiaddr==0.0.9
multidict==5.0.2
netaddr==0.8.0
packaging==20.4
parsimonious==0.8.1
path==15.0.0
path.py==12.5.0
pexpect==4.8.0
pika==1.1.0
pluggy==0.13.1
protobuf==3.14.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pyaes==1.6.1
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
pyflakes==2.2.0
pyparsing==2.4.7
Pypubsub==4.0.3
pyrsistent==0.17.3
pytest==6.1.2
python-binance-chain==0.1.20
python-bitcoinrpc==1.0
python-dateutil==2.8.1
pytz==2020.4
pywallet==0.1.0
PyYAML==5.3.1
requests==2.25.0
rlp==2.0.0
rusty-rlp==0.1.15
secp256k1==0.13.2
sha256==0.1
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.4.1
tabulate==0.8.7
toml==0.10.2
toolz==0.11.1
two1==3.10.9
typing-extensions==3.7.4.3
ujson==4.0.1
urllib3==1.26.2
varint==1.0.2
web3==5.13.0
websocket-client==0.57.0
websockets==8.1
yarl==1.6.3
